Add test and concrete implementation for tax applier (BarcodeTaxApplier should be splitted into different functionalities since doesn't met the SRP).

+++ src/main/java/com/test/pos/Display.java
@@ -2,7 +2,7 @@ package com.test.pos;
 
 public interface Display {
 
-	void print(Long price);
+	void print(int price);
 
 	void print(String string);
